Fix error File not found
When I am using the `--src-tables-where-clauses` parameter I alltime getting error `File not found`.
In the instruction I see `Comma-separated list of WHERE clauses for tables to be migrated. Ex.: db1.t1="a=1 and b=2",db2.t2="c=3 and k=4"` but in the code it all time try open some file.